requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
google-api-python-client==2.150.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httpx==0.27.0
aiosqlite==0.20.0
pyyaml==6.0.2
python-dotenv==1.0.1
apscheduler==3.10.4
gspread==6.1.2
pytest==8.3.0
pytest-asyncio==0.24.0
